cent_amount  #931
git-svn-id: file:///data/svn/enum/server/trunk@17358 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/fredlib/invoicing/invoice.h
@@ -20,7 +20,7 @@ namespace Invoicing {
 
 // input is price in cents(penny)
 // signed type allows negative amounts
-typedef signed int cent_amount;
+typedef unsigned long cent_amount;
 
 cent_amount get_price(const std::string &str);
